requirements.txt
Django==1.8.13
django-extensions==1.6.7
djangorestframework==3.4.0
psycopg2==2.6.2
six==1.10.0
